extract SingleGrill to new directory. update import path in PrimaryLayout
